requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
firebase-admin==6.3.0
google-cloud-firestore==2.16.0
google-cloud-storage==2.14.0
cryptography==42.0.0
Pillow==10.3.0
numpy==1.26.0
opencv-python==4.10.0.84
qrcode==7.4.2
APScheduler==3.11.0
gunicorn==21.2.0
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
ultralytics==8.3.228
transformers==4.47.1
huggingface-hub==0.34.3
tokenizers==0.21.2
safetensors==0.5.3
easyocr==1.7.1
pytesseract==0.3.13
deepface==0.0.79
tensorflow==2.19.1
